// src/@types/interfaces/IPayment.ts
export interface IPayment {
  id?: string;
  code?: string;
  branchId: string;
  branch?: string;
  employeeId?: string;
  employee?: string;
  amount?: number;
  date?: Date;
  note?: string;
  fromAccountId?: string;
  toAccountId?: string;
  fromAccount?: string;
  toAccount?: string;
  dateTo?: Date;
  dateFrom?: Date;
}